rename: email 추가로 인한 기존 클래스명 변경 1. myCar 큐 클래스명 추가
