+++ src/main/java/com/example/osid/event/OrderCompletedMyCarEvent.java
@@ -11,15 +11,15 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class OrderCompletedEvent implements Serializable {
+public class OrderCompletedMyCarEvent implements Serializable {
 
 	// orderId, (실패했을 경우)에러 메세지, 재시도 횟수
 	private Long orderId;
 	private String errorMessage;
 	private int retryCount;
 
-	public OrderCompletedEvent(Long orderId) {
+	public OrderCompletedMyCarEvent(Long orderId) {
 		this.orderId = orderId;
 	}
-	
+
 }
