+++ src/main/java/com/example/osid/event/listener/OrderCompletedMyCarListener.java
@@ -1,4 +1,4 @@
-package com.example.osid.event;
+package com.example.osid.event.listener;
 
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
@@ -9,6 +9,8 @@ import com.example.osid.config.RabbitMQConfig;
 import com.example.osid.domain.mycar.exception.MyCarException;
 import com.example.osid.domain.mycar.service.MyCarService;
 import com.example.osid.domain.order.exception.OrderException;
+import com.example.osid.event.OrderCompletedEmailEvent;
+import com.example.osid.event.OrderCompletedMyCarEvent;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -17,7 +19,7 @@ import lombok.extern.slf4j.Slf4j;
 @Component
 @RequiredArgsConstructor
 @ConditionalOnProperty(name = "mq.enabled", havingValue = "true", matchIfMissing = false)
-public class OrderCompletedListener {
+public class OrderCompletedMyCarListener {
 
 	// 최대 재시도 횟수
 	private static final int MAX_RETRY = 3;
@@ -25,10 +27,11 @@ public class OrderCompletedListener {
 	private final MyCarService myCarService;
 	private final RabbitTemplate rabbitTemplate;
 
-	@RabbitListener(queues = RabbitMQConfig.ORDER_COMPLETE_QUEUE)
-	public void handleOrderCompleted(OrderCompletedEvent event) {
+	// myCar 생성 -> email 발행
+	@RabbitListener(queues = RabbitMQConfig.MY_CAR_QUEUE)
+	public void handleOrderCompleted(OrderCompletedMyCarEvent event) {
 		int retryCount = event.getRetryCount();
-		log.info("이벤트 수신: orderId={}, retryCount={}", event.getOrderId(), retryCount);
+		log.info("myCar 생성 이벤트 수신: orderId={}, retryCount={}", event.getOrderId(), retryCount);
 
 		try {
 			myCarService.saveMyCar(event.getOrderId());
@@ -54,18 +57,24 @@ public class OrderCompletedListener {
 		}
 	}
 
-	private void resendToQueue(OrderCompletedEvent event) {
+	private void resendToQueue(OrderCompletedMyCarEvent event) {
 		log.info("재시도 메시지 전송: orderId={}, retryCount={}", event.getOrderId(), event.getRetryCount());
 		rabbitTemplate.convertAndSend(RabbitMQConfig.EXCHANGE, RabbitMQConfig.ROUTING_KEY, event);
 	}
 
-	private void sendToDlq(OrderCompletedEvent event) {
+	private void sendToDlq(OrderCompletedMyCarEvent event) {
 		log.error("DLQ 전송 : orderId={}, error={}", event.getOrderId(), event.getErrorMessage());
-		OrderCompletedEvent toSend = new OrderCompletedEvent(
+		OrderCompletedMyCarEvent toSend = new OrderCompletedMyCarEvent(
 			event.getOrderId(),
 			event.getErrorMessage(),
 			event.getRetryCount()
 		);
-		rabbitTemplate.convertAndSend(RabbitMQConfig.DLX_EXCHANGE, RabbitMQConfig.DLQ_ROUTING_KEY, toSend);
+		rabbitTemplate.convertAndSend(RabbitMQConfig.MY_CAR_DLX, RabbitMQConfig.MY_CAR_DLQ_ROUTING_KEY, toSend);
+	}
+
+	private void publishEmailEvent(Long orderId) {
+		OrderCompletedEmailEvent emailEvent = new OrderCompletedEmailEvent(orderId);
+		rabbitTemplate.convertAndSend(RabbitMQConfig.EXCHANGE, RabbitMQConfig.ROUTING_KEY, emailEvent);
+		log.info("이메일 전송 성공: orderId={}", orderId);
 	}
 }
